Remove some styling options
"No Underline on Sauce"
This should be a non-issue, as any styling scripts should deal with it
accordingly.
If not, any 4chan scripts provides options for custom CSS.

"Lowercase on 4chan"
Why not just make it lowercase everywhere?
Furthermore, this can be accomplished using CSS, which also allows
custom CSS to change it.